# Rewrite UTF-8 validation in shift-based DFA for 53%~133% performance increase on non-ASCII strings