@the8472 2025-02-07
> or 2KiB rodata.

I suspect that this is going to eat into cache capacity, especially in code that interleaves string validation with other things, e.g. CBOR decoding. Have you benchmarked this for short inputs and with mixed workloads?

In coretests we have [TINY/SMALL/MEDIUM/LARGE/HUGE](https://github.com/rust-lang/rust/blob/master/library/coretests/benches/str/corpora.rs) for that reason