@thaliaarchi 2025-02-07
> Should we also replace the implementation of [lossy conversion](https://github.com/rust-lang/rust/blob/c0639b8cad126d886ddd88964f729dd33fb90e67/library/core/src/str/lossy.rs#L194) by calling the new validation function?
> It has a very similar code doing almost the same thing.

FYI, in https://github.com/rust-lang/rust/pull/136677, I fixed the handling of invalid UTF-8 in `Display` for `OsStr`/`Path`, so there's a new user of lossy conversion. It needs to count the number of (potentially invalid) characters in a byte string and truncate it to some character width. I [extracted](https://github.com/rust-lang/rust/pull/136677/commits/d53ca9111f3140e05453d6fe59e49d76e8d20b70) a single iteration of the loop in `Utf8Chunks::next` to reuse it there. If `Utf8Chunks` is more amenable to using your optimization, `OsStr` might need a separate API.