@oxalica 2025-02-07
@the8472 
> > or 2KiB rodata.
> 
> I suspect that this is going to eat into cache capacity, especially in code that interleaves string validation with other things, e.g. CBOR decoding. Have you benchmarked this for short inputs and with mixed workloads?

Could you please explain how to test "short inputs and with mixed workloads"? Should I bench the performance of a CBOR library like [ciborium](https://github.com/enarx/ciborium) using the new std on this branch? I'll give it a check later.

> In coretests we have [TINY/SMALL/MEDIUM/LARGE/HUGE](https://github.com/rust-lang/rust/blob/master/library/coretests/benches/str/corpora.rs) for that reason

It seems they are only used in little places like `str::char_count` and `str::debug`. `str::char()` is untouchedby this PR, because it uses the [UTF-8 decoder `next_code_point`](https://github.com/rust-lang/rust/blob/01e4f19cc8027925ffe0885a86388b700e46bfab/library/core/src/str/validations.rs#L36), not the validator. `str as Debug` is unrelated to validation either. This PR is only changing the implementation of `core::str::from_utf8` and its derivations. 

> Additionally you may want to benchmark 4-byte characters such as emoji or maybe some exotic scripts on the supplementary plane.

It should performs the same as the "zh" benchmark. Since the DFA is agnostic to the input, as long as it's purely non-ASCII to get into the DFA path, the performance should be the same. The correctness of DFA can be audited by checking state transitions table to match RFC3629.
Nevertheless, here's the benchmark (`./x.py bench --stage 0 library/coretests -- str::str_validate_emoji`):

```
Before:
str::str_validate_emoji                                         2742.95ns/iter   +/- 46.44
After:
str::str_validate_emoji                                         1529.48ns/iter    +/- 7.52
```