@hanna-kruppe 2025-02-07
> No worries on my part as the current assignee -- in fact, if another reviewer has more context here, I'd be happy to yield this one.

I just meant that I, personally, couldn't take over the review even if I wanted to since I haven't had bors permissions in years 😄