@hanna-kruppe 2025-02-07
Slightly silly idea, free to a good home: mapping the input bytes into equivalence classes before feeding them to the DFA, as done in [Flexible and Economical UTF-8 Decoder](https://bjoern.hoehrmann.de/utf-8/decoder/dfa/), would reduce the rodata size to 256B + 8B * (# classes). The critical path for the DFA transition is unaffected, so this would still run pretty fast. On Skylake-ish hardware, idealized peak throughput goes from 1 cycle per byte to 1.5cpb because the extra load becomes the bottleneck (that's for a loop that does nothing but compute the final state, the version in this PR doesn't seem to hit 1cpb anyway). On CPUs that can sustain more than two loads per cycle, it may not even be any slower for long inputs.