@the8472 2025-02-07
> It seems they are only used in little places like str::char_count and str::debug. str::char() is untouchedby this PR

Sorry, I didn't mean that this was specifically for utf8 validation, I was just gesturing at more cases that can be useful for string benchmarking in general to cover common scenarios.
Maybe you can look at earlier PRs that touched validation, sometimes other people have implemented their own external benchmark suites if I recall correctly.

> Could you please explain how to test "short inputs and with mixed workloads"? Should I bench the performance of a CBOR library like [ciborium](https://github.com/enarx/ciborium) using the new std on this branch? I'll give it a check later.

That's the tricky part. As hanna-kruppe says, such effects are most visible in complex codebases and are difficult to test with micro-benchmarks.

I suggested CBOR because it should contain short strings (e.g. dictionary keys) mixed with other non-string parsing and some implementations might already have their own benchmarks that can be reused. So it could be a benchmark of intermediate complexity.
You can try using `perf stat` to see if an existing benchmark is already experiencing cache pressure which might be exacerbated by the lookup table. Though I guess modern CPU caches may have some strategy that can deal with streaming workloads without clobbering the parts, so maybe it won't be complex enough.

I'll kick off a perf run to see if it impacts rustc itself. But we have already eliminated a bunch of string validations in the compiler, so it likely isn't a good benchmark either.

@bors try @rust-timer queue

>  Nevertheless, here's the benchmark (./x.py bench --stage 0 library/coretests -- str::str_validate_emoji):

👍