@oxalica 2025-02-07
@hanna-kruppe
> Slightly silly idea, free to a good home: mapping the input bytes into equivalence classes before feeding them to the DFA, as done in [Flexible and Economical UTF-8 Decoder](https://bjoern.hoehrmann.de/utf-8/decoder/dfa/), would reduce the rodata size to 256B + 8B * (# classes). That's 256B + 96B with the classes from the linked implementation, but idk if that number can be lower or has to be larger in this context.
> The critical path for the DFA transition is not affected by the input remapping

Not really. As I mentioned in the 32-bit-shift part, the bottleneck of the DFA path is latency, since the DFA has data dependency from each state to its precedent state. The later state can only be calculated after the previous result is out, no matter how many ALUs you have. So,

> goes from 1 cycle per byte to 1.5cpb

... literally means +50% latency and -50% performance. I tested it by introducing an identity lookup table `[u8; 256]` before each `TRANS_TABLE` lookup, then `throughput-zh-aligned/shift-dfa-m16-a16` goes down from 3.401GiB/s to 2.424GiB/s, which is -29%. So I don't think it worth it.

But yes, the table must be fully in cache to achieve these numbers, mixed workload is a challenge. I also prefer smaller table size, but not yet find a way to shrink it without introducing massive performance drop.